+++ src/components/shared/CostDisplay.tsx
@@ -1,7 +1,7 @@
 import { FunctionalComponent, h } from '@stencil/core';
-import { $ } from '../../../utils/currency';
-import { meteredFeatureDisplayValue } from '../../../utils/plan';
-import { PlanMeteredFeatureEdge } from '../../../types/graphql';
+import { $ } from '../../utils/currency';
+import { meteredFeatureDisplayValue } from '../../utils/plan';
+import { PlanMeteredFeatureEdge } from '../../types/graphql';
 
 /**
  * Base Cost
